+++ Additional_src/Flye/src/repeat_graph/main_repeat.cpp
@@ -190,7 +190,7 @@ int repeat_main(int argc, char** argv)
 	Logger::get().info() << "Building repeat graph";
 	SequenceContainer edgeSequences;
 	RepeatGraph rg(seqAssembly, &edgeSequences);
-	rg.build();
+	rg.build(keepHaplotypes);
 	//rg.validateGraph();
 
 	Logger::get().info() << "Parsing reads";
@@ -261,7 +261,7 @@ int repeat_main(int argc, char** argv)
 		Logger::get().debug() << "[SIMPL] == Iteration " << iterNum << " ==";
 
 		actions += multInf.splitNodes();
-		if (isMeta) 
+		if (isMeta && !keepHaplotypes) 
 		{
 			actions += multInf.disconnectMinorPaths();
 		}
@@ -277,7 +277,7 @@ int repeat_main(int argc, char** argv)
 		if (!actions) break;
 	}
 
-	if (isMeta) 
+	if (isMeta && !keepHaplotypes) 
 	{
 		multInf.resolveForks();
 	}