+++ Additional_src/Flye/src/repeat_graph/repeat_graph.h
@@ -250,7 +250,7 @@ public:
 	{}
 	~RepeatGraph();
 
-	void build();
+	void build(bool keepHaplotypes);
 	void updateEdgeSequences();
 	void storeGraph(const std::string& filename);
 	void loadGraph(const std::string& filename);