+++ Additional_src/Flye/src/sequence/sequence_container.h
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <cstdint>
 #include <vector>
 #include <unordered_map>
 #include <string>