+++ Additional_src/Modified_hifiasm/Assembly.cpp
@@ -649,13 +649,6 @@ static void worker_ovec_cal0(void *data, long i, int tid)
 {
 	ha_ovec_buf_t *b = ((ha_ovec_buf_t**)data)[tid];
 	int fully_cov, abnormal;
-    // if(i != 33) return;
-    // fprintf(stderr, "[M::%s-beg] rid->%ld\n", __func__, i);
-    // if (memcmp("m64012_190920_173625/88015004/ccs", Get_NAME((R_INF), i), Get_NAME_LENGTH((R_INF),i)) == 0) {
-    //     fprintf(stderr, "[M::%s-beg] rid->%ld\n", __func__, i);
-    // } else {
-    //     return;
-    // }
 
     ha_get_candidates_interface(b->ab, i, &b->self_read, &b->olist, &b->olist_hp, &b->clist, 
     0.02, asm_opt.max_n_chain, 1, NULL/**&(b->k_flag)**/, &b->r_buf, &(R_INF.paf[i]), &(R_INF.reverse_paf[i]), &(b->tmp_region), NULL, &(b->sp));
@@ -664,17 +657,24 @@ static void worker_ovec_cal0(void *data, long i, int tid)
 	clear_Round2_alignment(&b->round2);
 
 	correct_overlap(&b->olist, &R_INF, &b->self_read, &b->correct, &b->ovlp_read, &b->POA_Graph, &b->DAGCon,
-			&b->cigar1, &b->hap, &b->round2, &b->r_buf, &(b->tmp_region.w_list), 0, 0, &fully_cov, &abnormal);
+			&b->cigar1, &b->hap, &b->round2, &b->r_buf, &(b->tmp_region.w_list), 0, 0/**1***/, &fully_cov, &abnormal);
 
 	b->num_read_base += b->self_read.length;
 	b->num_correct_base += b->correct.corrected_base;
 	b->num_recorrect_base += b->round2.dumy.corrected_base;
 
-	
 
-	R_INF.paf[i].is_fully_corrected = 0;
-	R_INF.paf[i].is_abnormal = abnormal;
+	// R_INF.paf[i].is_fully_corrected = 0;
+	// R_INF.paf[i].is_abnormal = abnormal;
     R_INF.trio_flag[i] = AMBIGU;
+
+    // R_INF.paf[i].is_fully_corrected = 0;
+    // if (fully_cov) {
+    //     if (get_cigar_errors(&b->cigar1) == 0 && get_cigar_errors(&b->round2.cigar) == 0)
+    //         R_INF.paf[i].is_fully_corrected = 1;
+    // }
+    // R_INF.paf[i].is_abnormal = abnormal;
+    // R_INF.trio_flag[i] = AMBIGU;
     
     push_final_overlaps(&(R_INF.paf[i]), R_INF.reverse_paf, &b->olist, 1);
     push_final_overlaps(&(R_INF.reverse_paf[i]), R_INF.reverse_paf, &b->olist, 2);
@@ -700,6 +700,7 @@ static void worker_ovec_related_reads(void *data, long i, int tid)
 
     if(k < R_INF_FLAG.query_num)
     {
+        R_INF_FLAG.read_id[k] = i;
         int fully_cov, abnormal, q_idx = k;
 
         ha_get_candidates_interface(b->ab, i, &b->self_read, &b->olist, &b->olist_hp, &b->clist, 
@@ -766,6 +767,7 @@ static void worker_ovec_related_reads(void *data, long i, int tid)
             Get_READ_LENGTH(R_INF, b->olist.list[k].y_id));
 		}
 
+        /**
         fprintf(R_INF_FLAG.fp, "***************************unmatched ovlp***************************\n");
         for (k = 0; k < b->olist.length; k++) 
         {
@@ -777,6 +779,7 @@ static void worker_ovec_related_reads(void *data, long i, int tid)
             b->olist.list[k].y_pos_strand, b->olist.list[k].strong, b->olist.list[k].without_large_indel,
             Get_READ_LENGTH(R_INF, b->olist.list[k].y_id));
 		}
+        **/
 
         R_INF.trio_flag[i] = AMBIGU;
 
@@ -917,7 +920,7 @@ void rescue_hp_reads(ha_ovec_buf_t **b)
     int hom_cov, het_cov;
     ha_flt_tab_hp = ha_idx_hp = NULL;
     if (!(asm_opt.flag & HA_F_NO_KMER_FLT)) {
-       ha_flt_tab_hp = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 1);
+       ha_flt_tab_hp = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 1, 0);
     }
     ha_idx_hp = ha_pt_gen(&asm_opt, ha_flt_tab, 1, 1, &R_INF, &hom_cov, &het_cov);
 
@@ -948,6 +951,19 @@ void print_het_cnt_log(uint32_t *het_cnt)
         fprintf(output_file, ">%.*s\t%u\n", (int)Get_NAME_LENGTH(R_INF, i), Get_NAME(R_INF, i), het_cnt[i]);
     }
     fclose(output_file);
+}
+
+void prt_dbg_rs(FILE *fp, Debug_reads* x, uint64_t round)
+{
+    uint64_t k, id; UC_Read g_read; init_UC_Read(&g_read);
+    for (k = 0; k < R_INF_FLAG.query_num; k++) {
+        id = x->read_id[k];
+        if(id == ((uint64_t)-1)) continue;
+        recover_UC_Read(&g_read, &R_INF, id);
+        fprintf(fp, ">%.*s_r%lu\n", (int)Get_NAME_LENGTH((R_INF), id), Get_NAME((R_INF), id), round);
+        fprintf(fp, "%.*s\n", (int)g_read.length, g_read.seq);
+    }
+    destory_UC_Read(&g_read);
 }
 
 
@@ -999,8 +1015,7 @@ void ha_overlap_and_correct(int round)
 		ha_ovec_destroy(b[i]);
 	}
 	free(b);
-
-	if (asm_opt.required_read_name) destory_Debug_reads(&R_INF_FLAG), exit(0); // for debugging only
+    if (asm_opt.required_read_name) prt_dbg_rs(R_INF_FLAG.fp_r0, &R_INF_FLAG, 0); // for debugging only
     
 	// save corrected reads to R_INF
 	CALLOC(e, asm_opt.thread_num);
@@ -1017,10 +1032,13 @@ void ha_overlap_and_correct(int round)
 		free(e[i].second_round_read);
 	}
 	free(e);
+
+    if (asm_opt.required_read_name) prt_dbg_rs(R_INF_FLAG.fp_r1, &R_INF_FLAG, 1); // for debugging only
+    if (asm_opt.required_read_name) destory_Debug_reads(&R_INF_FLAG), exit(0); // for debugging only
     ///debug_print_pob_regions();
 }
 
-void ha_overlap_cal(int round)
+void ha_overlap_cal(int round, int read_from_store)
 {
 	int i, hom_cov, het_cov;
 	ha_ovec_buf_t **b;
@@ -1031,9 +1049,9 @@ void ha_overlap_cal(int round)
 	for (i = 0; i < asm_opt.thread_num; ++i)
 		b[i] = ha_ovec_init(0, (round == asm_opt.number_of_round - 1),0);
     if(ha_idx) hom_cov = asm_opt.hom_cov;
-	if(ha_idx == NULL) ha_idx = ha_pt_gen(&asm_opt, ha_flt_tab, round == 0? 0 : 1, 0, &R_INF, &hom_cov, &het_cov); // build the index
+	if(ha_idx == NULL) ha_idx = ha_pt_gen(&asm_opt, ha_flt_tab, ((round == 0)&&(read_from_store == 0))?0:1, 0, &R_INF, &hom_cov, &het_cov); // build the index
 	///debug_adapter(&asm_opt, &R_INF);
-    if (round == 0 && ha_flt_tab == 0) // then asm_opt.hom_cov hasn't been updated
+    if (/**round == 0 &&**/ ha_flt_tab == 0) // then asm_opt.hom_cov hasn't been updated
 		ha_opt_update_cov(&asm_opt, hom_cov);
     het_cnt = NULL;
     if(round == asm_opt.number_of_round-1 && asm_opt.is_dbg_het_cnt) CALLOC(het_cnt, R_INF.total_reads);
@@ -1057,6 +1075,8 @@ void ha_overlap_cal(int round)
 		ha_ovec_destroy(b[i]);
 	}
 	free(b);
+    asm_opt.hom_cov = hom_cov;
+    asm_opt.het_cov = het_cov;
 }
 
 
@@ -1769,7 +1789,7 @@ void hap_recalculate_peaks(char* output_file_name)
     int hom_cov, het_cov;
     // construct hash table for high occurrence k-mers
     if (!(asm_opt.flag & HA_F_NO_KMER_FLT)) {
-        ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0);
+        ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0, 0);
         ha_opt_update_cov(&asm_opt, hom_cov);
     }
     free(R_INF.read_length);
@@ -1880,13 +1900,13 @@ int ha_assemble_ovec(void)
     // construct hash table for high occurrence k-mers
     if (!(asm_opt.flag & HA_F_NO_KMER_FLT) && ha_flt_tab == NULL) 
     {
-        ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0);
+        ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0, 0);
         ha_opt_update_cov(&asm_opt, hom_cov);
     }
     // error correction
     assert(asm_opt.number_of_round > 0);
     ha_opt_reset_to_round(&asm_opt, r); // this update asm_opt.roundID and a few other fields
-    ha_overlap_cal(r);
+    ha_overlap_cal(r, 0);
     fprintf(stderr, "[M::%s] size of buffer: %.3fGB\n", __func__, asm_opt.mem_buf / 1073741824.0);
     
     
@@ -1926,7 +1946,7 @@ int ha_assemble(void)
 		// construct hash table for high occurrence k-mers
 		if (!(asm_opt.flag & HA_F_NO_KMER_FLT) && ha_flt_tab == NULL) 
         {
-			ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0);
+			ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0, 0);
 			ha_opt_update_cov(&asm_opt, hom_cov);
 		}
 		// error correction
@@ -1950,6 +1970,72 @@ int ha_assemble(void)
 		ha_ft_destroy(ha_flt_tab);
 		if (asm_opt.flag & HA_F_WRITE_PAF) Output_PAF();
 		ha_triobin(&asm_opt);
+	}
+    if(ovlp_loaded == 2) ovlp_loaded = 0;
+    ha_opt_update_cov_min(&asm_opt, asm_opt.hom_cov, MIN_N_CHAIN);
+
+    build_string_graph_without_clean(asm_opt.min_overlap_coverage, R_INF.paf, R_INF.reverse_paf, 
+        R_INF.total_reads, R_INF.read_length, asm_opt.min_overlap_Len, asm_opt.max_hang_Len, asm_opt.clean_round, 
+        asm_opt.gap_fuzz, asm_opt.min_drop_rate, asm_opt.max_drop_rate, asm_opt.output_file_name, asm_opt.large_pop_bubble_size, 0, !ovlp_loaded);
+	destory_All_reads(&R_INF);
+	return 0;
+}
+
+
+int ha_assemble_pair(void)
+{
+	extern void ha_extract_print_list(const All_reads *rs, int n_rounds, const char *o);
+	int r = 0, hom_cov = -1, ovlp_loaded = 0; memset((&R_INF), 0, sizeof(R_INF));
+
+	if (asm_opt.load_index_from_disk && load_all_data_from_disk(&R_INF.paf, &R_INF.reverse_paf, asm_opt.output_file_name)) {
+		ovlp_loaded = 1;
+		fprintf(stderr, "[M::%s::%.3f*%.2f] ==> loaded corrected reads and overlaps from disk\n", __func__, yak_realtime(), yak_cpu_usage());
+		if (asm_opt.extract_list) {
+			ha_extract_print_list(&R_INF, asm_opt.extract_iter, asm_opt.extract_list);
+			exit(0);
+		}
+		if (asm_opt.flag & HA_F_WRITE_EC) Output_corrected_reads();
+		if (asm_opt.flag & HA_F_WRITE_PAF) Output_PAF();
+        if (asm_opt.het_cov == -1024) hap_recalculate_peaks(asm_opt.output_file_name), ovlp_loaded = 2;
+	}
+
+
+	if (!ovlp_loaded) {
+        
+        if(!append_All_reads(&R_INF, asm_opt.output_file_name, 0)) {
+            fprintf(stderr, "[M::%s::] Cannot load %s.0\n", __func__, asm_opt.output_file_name); 
+            exit(1);
+        }
+
+        if(!append_All_reads(&R_INF, asm_opt.output_file_name, 1)) {
+            fprintf(stderr, "[M::%s::] Cannot load %s.1\n", __func__, asm_opt.output_file_name); 
+            exit(1);
+        }
+        // Output_corrected_reads(); exit(0);
+
+        ha_flt_tab = ha_idx = NULL; r = asm_opt.number_of_round - 1;
+        if((asm_opt.flag & HA_F_VERBOSE_GFA)) load_pt_index(&ha_flt_tab, &ha_idx, &R_INF, &asm_opt, asm_opt.output_file_name), load_ct_index(&ha_ct_table, asm_opt.output_file_name);
+
+        // construct hash table for high occurrence k-mers
+        if (!(asm_opt.flag & HA_F_NO_KMER_FLT) && ha_flt_tab == NULL) {
+            ha_flt_tab = ha_ft_gen(&asm_opt, &R_INF, &hom_cov, 0, 1);
+            ha_opt_update_cov(&asm_opt, hom_cov);
+        }
+        // error correction
+        assert(asm_opt.number_of_round > 0);
+        ha_opt_reset_to_round(&asm_opt, r); // this update asm_opt.roundID and a few other fields
+        ha_overlap_cal(r, 1);
+        fprintf(stderr, "[M::%s] size of buffer: %.3fGB\n", __func__, asm_opt.mem_buf / 1073741824.0);
+        fprintf(stderr, "[M::%s::%.3f*%.2f@%.3fGB] ==> found overlaps for the final round\n", __func__, yak_realtime(),
+				yak_cpu_usage(), yak_peakrss_in_gb());
+        
+
+        if (asm_opt.flag & HA_F_WRITE_EC) Output_corrected_reads();
+        ha_print_ovlp_stat(R_INF.paf, R_INF.reverse_paf, R_INF.total_reads);
+        ha_ft_destroy(ha_flt_tab);
+        if (asm_opt.flag & HA_F_WRITE_PAF) Output_PAF();
+        ha_triobin(&asm_opt);
+
 	}
     if(ovlp_loaded == 2) ovlp_loaded = 0;
     ha_opt_update_cov_min(&asm_opt, asm_opt.hom_cov, MIN_N_CHAIN);