+++ Additional_src/Modified_hifiasm/Process_Read.h
@@ -25,6 +25,7 @@
 #define Get_NAME(R_INF, ID) ((R_INF).name + (R_INF).name_index[(ID)])
 #define CHECK_BY_NAME(R_INF, NAME, ID) (Get_NAME_LENGTH((R_INF),(ID))==strlen((NAME)) && \
                                         memcmp((NAME), Get_NAME((R_INF), (ID)), Get_NAME_LENGTH((R_INF),(ID))) == 0)
+#define IS_SCAF_READ(R_INF, ID) ((R_INF).read_sperate[(ID)] == NULL)
 
 extern uint8_t seq_nt6_table[256];
 extern char bit_t_seq_table[256][4];
@@ -151,9 +152,10 @@ typedef struct
 typedef struct
 {
     char** read_name;
+    uint64_t *read_id;
     uint64_t query_num;
     kvec_t_u64_warp* candidate_count;
-    FILE* fp;
+    FILE *fp, *fp_r0, *fp_r1;
     pthread_mutex_t OutputMutex;
 } Debug_reads;
 
@@ -203,6 +205,13 @@ typedef struct
     // uint32_t mm;
 } all_ul_t;
 
+
+typedef struct {
+    ul_vec_t *a;
+    size_t n, m;
+    uint8_t dd;
+} scaf_res_t;
+
 extern all_ul_t UL_INF;
 extern all_ul_t ULG_INF;
 // extern uint32_t *het_cnt;
@@ -220,6 +229,7 @@ void destory_UC_Read(UC_Read* r);
 void reverse_complement(char* pattern, uint64_t length);
 void write_All_reads(All_reads* r, char* read_file_name);
 int load_All_reads(All_reads* r, char* read_file_name);
+int append_All_reads(All_reads* r, char *idx, uint32_t id);
 void destory_All_reads(All_reads* r);
 int destory_read_bin(All_reads* r);
 void init_Debug_reads(Debug_reads* x, const char* file);
@@ -238,5 +248,8 @@ uint64_t retrieve_r_cov_region(const ul_idx_t *ul, uint64_t id, uint8_t strand, 
 void append_ul_t_back(all_ul_t *x, uint64_t *rid, char* id, int64_t id_l, char* str, int64_t str_l, ul_ov_t *o, int64_t on, float p_chain_rate);
 void write_compress_base_disk(FILE *fp, uint64_t ul_rid, char *str, uint32_t len, ul_vec_t *buf);
 int64_t load_compress_base_disk(FILE *fp, uint64_t *ul_rid, char *dest, uint32_t *len, ul_vec_t *buf);
+scaf_res_t *init_scaf_res_t(uint32_t n);
+void destroy_scaf_res_t(scaf_res_t *p);
+void read_ma(ma_hit_t* x, FILE* fp);
 
 #endif