+++ Additional_src/Modified_hifiasm/horder.cpp
@@ -3816,13 +3816,13 @@ float chimeric_rate, float drop_ratio, int max_hang, int min_ovlp)
     ma_ug_seq_scaffold(ug, sg, coverage_cut, sources, &new_rtg_edges, max_hang, min_ovlp, 0, 1);
     ma_ug_print(ug, sg, coverage_cut, sources, ruIndex, "stg", output_file);
     fclose(output_file);
-	
-	if(asm_opt.make_only_primary_contigs == 1)
-	{
-		fprintf(stderr, "Modified_hifiasm finishes because the user has provided the option \"--only-primary\"");
-		exit(0);
-	}
-
+    
+    if(asm_opt.make_only_primary_contigs == 1)
+    {
+        fprintf(stderr, "Modified_hifiasm finishes because the user has provided the option \"--only-primary\"");
+        exit(0);
+    }
+    
     sprintf(gfa_name, "%s.%s.p_ctg.noseq.gfa", output_file_name, "stg");
     output_file = fopen(gfa_name, "w");
     ma_ug_print_simple(ug, sg, coverage_cut, sources, ruIndex, "stg", output_file);
@@ -3975,8 +3975,8 @@ asg_t *i_rg, ma_ug_t* i_ug, bubble_type* bub, kv_u_trans_t *ref, ug_opt_t *opt, 
 
 int cmp_mc_edge_t_w(const void * a, const void * b)
 {
-    if((*(osg_arc_t*)a).nw == (*(osg_arc_t*)b).nw) return 0;
-    return (*(osg_arc_t*)a).nw < (*(osg_arc_t*)b).nw ? -1 : 1;
+    if((*(mc_edge_t*)a).w == (*(mc_edge_t*)b).w) return 0;
+    return (*(mc_edge_t*)a).w < (*(mc_edge_t*)b).w ? -1 : 1;
 }
 
 void cal_chain_arch(scg_t *sg, const mc_match_t *ma, uint32_t v, uint32_t *va, uint32_t vn, uint64_t *idx, asg64_v *srt)
@@ -4309,6 +4309,7 @@ double min_cut, double max_cut, uint64_t cut_round)
         kv_pushp(mc_edge_t, sm, &sp);
         sp->w = sg->g->arc[k].nw; sp->x = k;
     }
+    // fprintf(stderr, "sm.n::%lu\n", (uint64_t)sm.n);
     qsort(sm.a, sm.n, sizeof(mc_edge_t), cmp_mc_edge_t_w);